move public space url copy buttons into invitetospace
